Refactoring Round 1
